bring origin back to u64
this caused an integer overflow crash, eventually we might want
to change this to u32 but the rest of the application must be
in a more mature state before this change can be done without
causing crashes

+++ src/awebo/Database/CommonQueries.zig
@@ -106,7 +106,7 @@ select_channel_messages: Query(
     .kind = .rows,
     .cols = struct {
         uid: u64,
-        origin: u32,
+        origin: u64,
         author: awebo.User.Id,
         body: []const u8,
     },
@@ -191,7 +191,7 @@ insert_message: Query(
     .kind = .exec,
     .args = struct {
         uid: u64,
-        origin: u32,
+        origin: u64,
         channel: awebo.Channel.Id,
         author: ?awebo.User.Id,
         body: []const u8,
